+++ SmashBcatDetector/Difference/Handlers/Discord/EventDiscordHandler.cs
@@ -15,7 +15,7 @@ namespace SmashBcatDetector.Difference.Handlers.Discord
 {
     public class EventDiscordHandler
     {
-        [DifferenceHandler(FileType.Event, DifferenceType.Added, 50)]
+        [DifferenceHandler((int)FileType.Event, DifferenceType.Added, 50)]
         public static async Task HandleAdded(Event addedEvent)
         {
             // Localize the embed description
